Added check to getImage for None
One of the comics (2198) isn't on the page as an img, in the same
way the other comics are. It's a series of interactive divs.

I added a quick check in getImage for comicImg returning None, and
it now only tries to dl/save the img if this is not None.